requirements/requirements-ingestion.txt
python-dotenv==1.0.1
kafka-python==2.3.1
requests==2.32.4
filelock>=3.13
